+++ opensquirrel/circuit_builder.py
@@ -9,7 +9,7 @@ from typing_extensions import Self
 
 from opensquirrel.circuit import Circuit
 from opensquirrel.default_instructions import default_instruction_set
-from opensquirrel.ir import IR, Bit, BitLike, Instruction, Qubit, QubitLike
+from opensquirrel.ir import IR, AsmDeclaration, Bit, BitLike, Instruction, Qubit, QubitLike
 from opensquirrel.register_manager import BitRegister, QubitRegister, RegisterManager
 
 
@@ -43,7 +43,7 @@ class CircuitBuilder:
         self.ir = IR()
 
     def __getattr__(self, attr: Any) -> Callable[..., Self]:
-        return partial(self._add_instruction, attr)
+        return partial(self._add_statement, attr)
 
     def _check_qubit_out_of_bounds_access(self, qubit: QubitLike) -> None:
         """Throw error if qubit index is outside the qubit register range.
@@ -73,11 +73,19 @@ class CircuitBuilder:
         for bit in instruction.get_bit_operands():
             self._check_bit_out_of_bounds_access(bit)
 
-    def _add_instruction(self, attr: str, *args: Any) -> Self:
+    def _add_statement(self, attr: str, *args: Any) -> Self:
+        if attr == "asm":
+            try:
+                asm_declaration = AsmDeclaration(*args)
+                self.ir.add_asm_declaration(asm_declaration)
+            except TypeError:
+                msg = f"trying to build '{attr}' with the wrong number or type of arguments: '{args}'"
+                raise TypeError(msg) from None
+            return self
+
         if attr not in default_instruction_set:
             msg = f"unknown instruction '{attr}'"
             raise ValueError(msg)
-
         try:
             instruction = default_instruction_set[attr](*args)
         except TypeError: