+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Assembly declaration
 - `CZDecomposer` decomposer pass
 - `ShortestPathRouter` router pass
 - `RandomMapper` mapper pass